Convert ProductPageBlock into RSC server component
