+++ apps/store/src/app/[locale]/StoryblokLayout.tsx
@@ -1,4 +1,3 @@
-'use client'
 import { clsx } from 'clsx'
 import type { ReactNode } from 'react'
 import { Fragment } from 'react'
@@ -8,7 +7,6 @@ import type { GlobalStory } from '@/services/storyblok/storyblok'
 import { filterByBlockType } from '@/services/storyblok/Storyblok.helpers'
 import { wrapper } from './StoryblokLayout.css'
 
-// TODO: Render layout and page server-side
 export const StoryblokLayout = ({
   children,
   globalStory,
@@ -18,8 +16,8 @@ export const StoryblokLayout = ({
   globalStory: GlobalStory
   className?: string
 }) => {
-  const headerBlock = filterByBlockType(globalStory.content.header, HeaderBlock.blockName)
-  const footerBlock = filterByBlockType(globalStory.content.footer, FooterBlock.blockName)
+  const headerBlock = filterByBlockType(globalStory.content.header, 'header')
+  const footerBlock = filterByBlockType(globalStory.content.footer, 'footer')
 
   return (
     <div className={clsx(wrapper, className)}>
