+++ apps/store/src/services/storyblok/commonStoryblokComponents.ts
@@ -35,7 +35,7 @@ import { PerilsBlock } from '@/blocks/PerilsBlock'
 import { ProductCardBlock } from '@/blocks/ProductCardBlock'
 import { ProductDocumentsBlock } from '@/blocks/ProductDocumentsBlock'
 import { ProductGridBlock } from '@/blocks/ProductGridBlock'
-import { ProductPageBlock } from '@/blocks/ProductPageBlock'
+import { ProductPageBlock } from '@/blocks/ProductPageBlock/ProductPageBlock'
 import { ProductPillowBlock } from '@/blocks/ProductPillowsBlock/ProductPillowBlock'
 import { ProductPillowsBlock } from '@/blocks/ProductPillowsBlock/ProductPillowsBlock'
 import { ProductReviewsBlock } from '@/blocks/ProductReviewsBlock'
